ThemePicker: Reimplement font picker
BP3A has enabled the new compose based picker UI, making the old
FontPicker invisible to ThemePicker.

This commit refactors the FontPicker to be compatible with the new
binders and uses compose to draw it on the screen. Some of the older
Java files that were still being used have been converted to kotlin.

The rewrite also changes how the picker is displayed on the screen.
The older implementation would open an entirely different section for
the font picker. However, the new ThemePicker expects the preview to be
either; 1) shown at all times inside a picker or 2) hidden but the
picker to be accessed inside Settings.

If we displayed the older UI on the new ThemePicker, it would not be
able to refresh the preview when we would exit the app and go back,
and would show a black screen. So the new implementation skips this and
draws a compact UI instead where the preview is visible.

The rewrite also adds a new scrolling method to longer strings to not
break the padding; `.basicMarquee()`.

Change-Id: I92420b0d098ebaa5b8245ed2911b5ade483384b9
Signed-off-by: sidharthify <wednisegit@gmail.com>
Signed-off-by: Ghosuto <clash.raja10@gmail.com>